Builder account creation code
All of the code required for the builder to create an account, by connecting to the server and giving them the builders email. for security passwords are hashed before being sent to the server so the server never actually knows the builders real password, only the encrypted one. also some very basic checks to make sure email is a valid email format.